[Breaking] Migrating a sorted table reverses it
When migrating a table that is sorted (default), the order of the
items within the table is reversed. This is not the intended behaviour
for migration, but it was not documented or tested. To avoid the case
where somebody is relying on this behaviour, I'm considering this bugfix
as including a breaking change and releasing it as a minor release.